Make authenticated staging probes initialize a fresh workspace
