+++ test/e2e/staging/lifecycle.spec.ts
@@ -1,4 +1,3 @@
-import { createCipheriv, createHash, randomBytes } from "node:crypto";
 import {
   type APIRequestContext,
   expect,
@@ -6,6 +5,7 @@ import {
   test
 } from "@playwright/test";
 import { stagingMailApiPath } from "./mail-api-path";
+import { ensureStagingSetup } from "./setup";
 
 const email = required("HQBASE_STAGING_OWNER_EMAIL");
 const password = required("HQBASE_STAGING_OWNER_PASSWORD");
@@ -48,34 +48,7 @@ test("HQBase web lifecycle remains healthy", async ({ page, request }) => {
     )
     .toBe(200);
 
-  const status = await request.get("/api/setup/status");
-  expect(status.ok()).toBeTruthy();
-  const setup = (await status.json()) as { isComplete: boolean };
-  if (!setup.isComplete) {
-    const grantCookie = stagingSetupGrantCookie(required("HQBASE_STAGING_AUTH_SECRET"));
-    const bootstrap = await request.post("/api/setup/bootstrap", {
-      data: {
-        checklistAcknowledged: true,
-        defaultFromMailboxAddress: sender,
-        emailDomains: [
-          {
-            catchAllMailboxAddress: sender,
-            catchAllPolicy: "mailbox",
-            name: domain
-          }
-        ],
-        mailboxes: [{ address: sender, displayName: "HQBase E2E" }],
-        ownerEmail: email,
-        ownerName: "HQBase E2E Owner",
-        ownerPassword: password,
-        portalHostname: new URL(stagingUrl).hostname,
-        primaryDomain: domain
-      },
-      headers: { cookie: grantCookie }
-    });
-    expect(bootstrap.status()).toBe(201);
-    await expect(bootstrap.json()).resolves.toMatchObject({ setup: { isComplete: true } });
-  }
+  await ensureStagingSetup(request);
 
   await expect
     .poll(async () => {
@@ -1007,18 +980,6 @@ function accessHeaders(): Record<string, string> {
         "CF-Access-Client-Secret": clientSecret
       }
     : { origin: stagingUrl };
-}
-
-function stagingSetupGrantCookie(secret: string): string {
-  const iv = randomBytes(12);
-  const key = createHash("sha256").update(`hqbase-runtime-cloudflare-oauth:${secret}`).digest();
-  const cipher = createCipheriv("aes-256-gcm", key, iv);
-  const encrypted = Buffer.concat([
-    cipher.update("hqbase-staging-oauth-grant", "utf8"),
-    cipher.final(),
-    cipher.getAuthTag()
-  ]);
-  return `hqb_cf_oauth_grant=${encodeURIComponent(`${iv.toString("base64url")}.${encrypted.toString("base64url")}`)}`;
 }
 
 function required(name: string): string {
